pyproject.toml
[project]
name = "archflow-hub"
version = "0.2.0"
description = "Context Hub MCP Server - Jira + GitHub + Draw.io unified intelligence"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [{ name = "Moobean Team" }]
keywords = ["mcp", "jira", "github", "drawio", "architecture", "context-hub"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "fastmcp>=2.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
]

[project.urls]
Homepage = "https://github.com/Juhwan01/ArchFlow"
Repository = "https://github.com/Juhwan01/ArchFlow"
Issues = "https://github.com/Juhwan01/ArchFlow/issues"

[project.scripts]
archflow = "archflow.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/archflow"]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
]
